server/src/main/java/com/example/server/controller/UserController.java
package com.example.server.controller;

import com.example.server.models.user.Role;
import com.example.server.models.user.User;
import com.example.server.services.UserService;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
public class UserController {

    private final UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }


    @GetMapping("/users")
    @PreAuthorize("hasRole('ADMIN')")
    public List<User> getAllUsers() {
        return userService.getAllUsers();
    }

    @PostMapping("/users")
    @PreAuthorize("hasRole('ADMIN')")

    public User addUser() {
        return userService.addUser(new User("lisa1", "password", Role.REGULAR_USER));
    }

    @DeleteMapping("/users/{id}")
    @PreAuthorize("hasRole('ADMIN')")

    public void deleteUser(@PathVariable Long id) {
        userService.deleteUserById(id);
    }

    @PutMapping("/users/{id}")
    @PreAuthorize("hasRole('ADMIN')")
    public void editUser(@PathVariable Long id, @RequestBody User newUser) {
        userService.editUser(id, newUser);
    }
}
